+++ inst/include/hesim/stats/survfuns.h
@@ -103,7 +103,7 @@ inline std::vector<double> cumhazard_numeric(Func hazfun, std::vector<double> ti
     std::vector<double> cumhazard(times.size());
     const double lower = 0;
     double abserr; int ier;
-    for (int i = 0; i < times.size(); ++i){
+    for (int i = 0; i < (int) times.size(); ++i){
       const double upper = times[i];
       cumhazard[i] = math::quad(hazfun, lower, upper, abserr, ier);
     }
@@ -153,7 +153,7 @@ inline double surv_sample(std::vector<double> &time, std::vector<double> cumhaz,
   double died = 0;
   int i = 1;
   unsigned int n_times = time.size();
-  while(died == 0 && i < n_times){
+  while(died == 0 && i < (int) n_times){
     double prob = 1 - exp(cumhaz[i - 1] - cumhaz[i]);
     died = rbernoulli(prob);
     if (died == 1){
